+++ src/DotNetRemoteWebDriver/Properties/AssemblyInfo.cs
@@ -37,6 +37,6 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 
-[assembly: AssemblyVersion("0.1.*")]
-[assembly: AssemblyFileVersion("0.1.0")]
+[assembly: AssemblyVersion("0.9.*")]
+[assembly: AssemblyFileVersion("0.9.0")]
 [assembly: AssemblyInformationalVersion("1.5.0")]