+++ src/DotNetRemoteWebDriver/ServiceProviderExtensons.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace DotNetRemoteWebDriver
+{
+    public static class ServiceProviderExtensons
+    {
+        public static TServiceType GetService<TServiceType>(this IServiceProvider serviceProvider)
+            where TServiceType : class
+        {
+            return (TServiceType) serviceProvider.GetService(typeof (TServiceType));
+        }
+    }
+}