+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Change Log
-## v0.1
 
+## v0.9
+- More tested out.
+- Kills created sub processes/drivers on exit.
+- Removed hard coded reference to firefox binary.
+- Tested with drivers and client api for selenium 2.53.
+
+## v0.1
 - Forked the Winium.Desktop project and built this initial version of the
 remote web driver.